refactor(evpn_originator): split mod.rs into concern submodules
Distribute the module-scope functions from the 5.8k-line root into concern
files: lifecycle, observation, duplicate_mac, rib_polling, rib_write, and tests.
Unlike peer_manager (impl methods), these are free functions, so cross-module
callers import them explicitly via crate::evpn_originator::<mod>::<fn>; siblings
use explicit imports (tests.rs uses sibling globs, allowed in test modules). All
moved fns are pub(super). mod.rs keeps state/structs/public impls + the spawn
fns + originator_loop (~845 lines). Pure relocation; 51 tests unchanged
(test-name receipt identical).
